+++ Source/Frontend/UI/Modular/UI_CanvasForm.cs
@@ -36,7 +36,7 @@
         {
             InitializeComponent();
 
-            UICore.SetRTCColor(UICore.GeneralColor, this);
+            Colors.SetRTCColor(Colors.GeneralColor, this);
 
             if (!extraForm)
             {
